update fiscal EsocialEventos

+++ backend/src/main/java/com/example/backend/fiscal/esocialEventos/EsocialEventosRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.fiscal.esocialEventos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EsocialEventosRepository extends JpaRepository<EsocialEventos, Integer> {
+public interface EsocialEventosRepository extends JpaRepository<EsocialEventos, Long> {
 }
